Move some job functions to common
Namely bytesToIntSlice, addToIntSlice, and intSliceToBytes
